+++ src/database/pagination.go
@@ -1,4 +1,4 @@
-package main
+package database
 
 import "math"
